✨ feat: ollama
