+++ examples/basic_ollama/simple_loop_no_helpers.py
@@ -1,4 +1,4 @@
-from swarm.core import Swarm, Agent
+from swarm_ollama.core import Swarm, Agent
 import ollama
 from typing import List, Dict, Any
 import json
@@ -42,8 +42,8 @@ class OllamaWrapper:
 
             return WrappedResponse(response)
 
-        def __getattr__(self, name):
-            return getattr(self.client, name)
+    def __getattr__(self, name):
+        return getattr(self.client, name)
 
 # Initialize Ollama client
 ollama_client = ollama.Client(host="http://localhost:11434")
@@ -54,48 +54,26 @@ wrapped_client = OllamaWrapper(ollama_client)
 # Initialize Swarm with wrapped client
 client = Swarm(client=wrapped_client)
 
-english_agent = Agent(
-    name="English Agent",
+my_agent = Agent(
     model="llama3.2:3b",
-    instructions="You only speak English.",
+    name="Agent",
+    instructions="You are a helpful agent.",
 )
 
-spanish_agent = Agent(
-    name="Spanish Agent",
-    model="llama3.2:3b",
-    instructions="You only speak Spanish.",
-)
-
-def transfer_to_spanish_agent():
-    """Transfer spanish speaking users immediately."""
-    return spanish_agent
-
-llama_tool_call = """[{
-        "name": "transfer_to_spanish_agent",
-        "description": "Transfer the conversation to a Spanish-speaking agent",
-        "parameters": {
-            "type": "object",
-            "properties": {}
-        }
-        ]"""
-
-english_agent.functions.append(transfer_to_spanish_agent)
-english_agent.instructions = f"""You only speak English. You can use the following functions:
-
-{llama_tool_call}
-
-If you need to use a function, format your response as:
-[function_name()]
-
-Only use the functions when necessary, and provide a natural language response after using them.
-If the user speaks Spanish or requests Spanish assistance, use the transfer_to_spanish_agent function."""
-
-messages = [{"role": "user", "content": "Hola. ¿Como estás?"}]
-response = client.run(agent=english_agent, messages=messages)
+def pretty_print_messages(messages):
+    for message in messages:
+        if message["content"] is None:
+            continue
+        print(f"{message['role']}: {message['content']}") #this was adopted to ollama style completion. 
 
 
-if "[transfer_to_spanish_agent(" in response.messages[-1]["content"]:
-    transfer_agent = transfer_to_spanish_agent()
-    result = client.run(agent=transfer_agent, messages=response.messages)
-    print(result.messages[-1]["content"])
-#print(response.messages[-1]["content"])
+messages = []
+agent = my_agent
+while True:
+    user_input = input("> ")
+    messages.append({"role": "user", "content": user_input})
+
+    response = client.run(agent=agent, messages=messages)
+    messages = response.messages
+    agent = response.agent
+    pretty_print_messages(messages)
