+++ examples/triage_agent/evals_util.py
@@ -13,7 +13,7 @@ class BoolEvalResult(BaseModel):
 
 def evaluate_with_llm_bool(instruction, data) -> BoolEvalResult:
     eval_result, _ = __client.chat.completions.create_with_completion(
-        model="gpt-4o",
+        model="llama3.2:3b",
         messages=[
             {"role": "system", "content": instruction},
             {"role": "user", "content": data},
