+++ examples/triage_agent/run.py
@@ -1,4 +1,4 @@
-from swarm.repl import run_demo_loop
+from swarm_ollama.repl import run_demo_loop
 from agents import triage_agent
 
 if __name__ == "__main__":
