+++ setup.cfg
@@ -1,8 +1,8 @@
 [metadata]
-name = swarm
+name = swarm-ollama
 version = 0.1.0
-author = OpenAI Solutions
-description = A lightweight, stateless multi-agent orchestration framework.
+author = Yash Gadbail, OpenAI Solutions
+description = A lightweight, stateless multi-agent orchestration framework with ollama.
 long_description = file: README.md
 long_description_content_type = text/markdown
 license = MIT
@@ -13,6 +13,7 @@ zip_safe = True
 include_package_data = True
 install_requires =
     numpy
+    ollama
     openai>=1.33.0
     pytest
     requests
