+++ swarm_ollama/util.py
