+++ tests/test_util.py
@@ -1,4 +1,4 @@
-from swarm.util import function_to_json
+from swarm_ollama.util import function_to_json
 
 
 def test_basic_function():
